adding a null check to destroy in UnityAudioEngineSource.cs (#429)
* adding a null check to destroy in UnityAudioEngineSource.cs

 mAudioSource can be nuked before the Steam Audio component. 
This handles that and allows Steam Audio to remove the source.

I would also add a debug logerror and potentially reorder this to do the Steam Audio remove source first. 

The fact that UpdateParameters checks this, but on destroy does not make me believe this was missed.

* Update UnityAudioEngineSource.cs

woops, you didnt see that.