[core] Fix issues with IPLMatrix4x4 layout
The API assumes that IPLMatrix4x4 is row-major, but ipl::Matrix4x4f is
column-major, and the API layer didn't apply the correct conversion. So
some code was applying the conversion in the wrong place, and other code
assumed that the caller applied the conversion, which was very
confusing.

This change makes it so all API functions that accept IPLMatrix4x4
convert from row-major to column-major before passing the matrix to
ipl::Matrix4x4f, and removes unnecessary conversion code in the Unity
and Unreal plugins.

See also: #427