[unreal] Fix for exporting instanced levels
This changes the behavior of the Export Static Geometry menu command to
ignore instanced levels, unless an instanced level is currently being
edited (by right-clicking an ALevelInstanceActor and choosing Edit), in
which case the menu command just exports static geometry for the
instanced level.

The intended workflow is to use the Steam Audio Dynamic Object component
in the instanced level, since this will correctly handle instancing and
transforms.

See also: #441